+++ src/TextUI/Application.php
@@ -54,7 +54,7 @@ use PHPUnit\Runner\Baseline\Writer;
 use PHPUnit\Runner\CodeCoverage;
 use PHPUnit\Runner\CodeCoverageInitializationStatus;
 use PHPUnit\Runner\DeprecationCollector\Facade as DeprecationCollector;
-use PHPUnit\Runner\DeprecationFilter\Filter;
+use PHPUnit\Runner\DeprecationFilter;
 use PHPUnit\Runner\DirectoryDoesNotExistException;
 use PHPUnit\Runner\ErrorHandler;
 use PHPUnit\Runner\Extension\ExtensionBootstrapper;
@@ -1002,12 +1002,12 @@ final readonly class Application
 
             $filter = new $className;
 
-            if (!$filter instanceof Filter) {
+            if (!$filter instanceof DeprecationFilter) {
                 EventFacade::emitter()->testRunnerTriggeredPhpunitWarning(
                     sprintf(
                         'Class %s cannot be used as a deprecation filter because it does not implement %s',
                         $className,
-                        Filter::class,
+                        DeprecationFilter::class,
                     ),
                 );
